+++ data/src/main/java/camp/nextstep/edu/github/data/repository/DefaultMainRepository.kt
@@ -2,8 +2,8 @@ package camp.nextstep.edu.github.data.repository
 
 import camp.nextstep.edu.github.data.service.GitHubService
 import camp.nextstep.edu.github.data.toDomain
-import camp.nextstep.edu.github.domain.GitHubRepos
-import camp.nextstep.edu.github.domain.MainRepository
+import camp.nextstep.edu.github.domain.model.GitHubRepos
+import camp.nextstep.edu.github.domain.repository.MainRepository
 import javax.inject.Inject
 
 internal class DefaultMainRepository @Inject constructor(
