--- domain/src/main/java/camp/nextstep/edu/github/domain/MainRepository.kt
@@ -1,5 +1,0 @@
-package camp.nextstep.edu.github.domain
-
-interface MainRepository {
-    suspend fun getGitHubRepositories() : GitHubRepos
-}
