+++ domain/src/main/java/camp/nextstep/edu/github/domain/model/GitHubRepo.kt
@@ -1,3 +1,3 @@
-package camp.nextstep.edu.github.domain
+package camp.nextstep.edu.github.domain.model
 
 data class GitHubRepo(val id : Long, val fullName: String, val description: String)
