+++ domain/src/main/java/camp/nextstep/edu/github/domain/model/GitHubRepos.kt
@@ -1,3 +1,3 @@
-package camp.nextstep.edu.github.domain
+package camp.nextstep.edu.github.domain.model
 
 data class GitHubRepos(val repositories: List<GitHubRepo>)
