+++ domain/src/main/java/camp/nextstep/edu/github/domain/repository/MainRepository.kt
@@ -1,0 +1,7 @@
+package camp.nextstep.edu.github.domain.repository
+
+import camp.nextstep.edu.github.domain.model.GitHubRepos
+
+interface MainRepository {
+    suspend fun getGitHubRepositories() : GitHubRepos
+}
